Cages
We also created an NFT and it was very nice and meaningful.  I would like to bring to your attention that Busexsen belongs to me and whoever gets it, of course, the powers will pass to him. Here is a photo about the human community.  It is as if we gather people in a cage and keep them in this state, like the borders of our state.  I thought a lot about the salty NFT tip and I see it as a first.  Salt is a simple thing, but it took me days to think about it. And I was very pleased with the work, and I loved it very much.